[CRES-110] refactor: http 메서드를 사용해 useQuery 커스텀 훅 명칭 변경 #158
